lzma-utils 4.32.6 from http://tukaani.org/lzma/download
git-svn-id: svn://tug.org/texlive/trunk/Build/source@8158 c570f23f-e606-0410-a88d-b1316a301751

+++ utils/lzma-utils/src/sdk/7zip/Compress/RangeCoder/RangeCoder.h
@@ -14,7 +14,7 @@ const UInt32 kTopValue = (1 << kNumTopBits);
 
 class CEncoder
 {
-  UInt32 _ffNum;
+  UInt64 _ffNum;
   Byte _cache;
 public:
   UInt64 Low;
